requirements.txt
torch
numpy
matplotlib
opencv-python
timm
einops
pyquaternion
nuscenes-devkit
jupyterlab
